script: Combine import key operations of RSA algorithms
The import key operations of

- RSASSA-PKCS1-v1_5 (`rsassa_pkcs1_v1_5_operation::import_key`)
- RSA-PSS (`rsa_pss::import_key`)
- RSA_OAEP (`rsa_oaep::import_key`)

only differ from each other by a few steps. This patch combines them
into a single function (`rsa_common::import_key`) shared among them.

Testing: Refactoring. Existing tests suffice.
